requirements.txt
streamlit
pandas
plotly
scikit-learn
python-dotenv
statsmodels
mysql-connector-python
sqlalchemy